+++ packages/backend/src/server/api/endpoints/i/import-muting.ts
@@ -40,7 +40,7 @@ export const meta = {
 	},
 } as const;
 
-const paramDef = {
+export const paramDef = {
 	type: 'object',
 	properties: {
 		fileId: { type: 'string', format: 'misskey:id' },