+++ packages/backend/src/server/api/endpoints/i/read-all-messaging-messages.ts
@@ -10,7 +10,7 @@ export const meta = {
 	kind: 'write:account',
 } as const;
 
-const paramDef = {
+export const paramDef = {
 	type: 'object',
 	properties: {},
 	required: [],