+++ packages/backend/src/server/api/endpoints/notes/children.ts
@@ -23,7 +23,7 @@ export const meta = {
 	},
 } as const;
 
-const paramDef = {
+export const paramDef = {
 	type: 'object',
 	properties: {
 		noteId: { type: 'string', format: 'misskey:id' },