+++ packages/backend/src/server/api/endpoints/server-info.ts
@@ -8,7 +8,7 @@ export const meta = {
 	tags: ['meta'],
 } as const;
 
-const paramDef = {
+export const paramDef = {
 	type: 'object',
 	properties: {},
 	required: [],