+++ packages/backend/src/server/api/endpoints/users/pages.ts
@@ -6,7 +6,7 @@ export const meta = {
 	tags: ['users', 'pages'],
 } as const;
 
-const paramDef = {
+export const paramDef = {
 	type: 'object',
 	properties: {
 		userId: { type: 'string', format: 'misskey:id' },